find_diff/requirements.txt
# requirements.txt
opencv-python==4.8.0.76
numpy==1.24.3
matplotlib==3.7.2
Pillow==10.0.0
scikit-image==0.21.0
TensorFlow==2.10
